fix(*)：提交未提交的部分

+++ src/main/java/com/sipc/mmtbackend/pojo/dto/result/dataDashboard/po/QuestionScorePo.java
@@ -14,6 +14,6 @@ public class QuestionScorePo {
 
     private String question;
 
-    private List<Integer> score;
+    private List<Double> score;
 
 }
